more linux stuff
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@250 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ simpleraytracer/raymesh.cpp
@@ -682,7 +682,7 @@ void RayMesh::computeShadingNormals()
 void RayMesh::mergeVerticesWithSamePosAndNormal()
 {
 	conPrint("Merging vertices for mesh '" + this->getName() + "'...");
-	conPrint("\tInitial num vertices: " + toString(vertices.size()));
+	conPrint("\tInitial num vertices: " + toString((unsigned int)vertices.size()));
 
 	std::map<RayMeshVertex, unsigned int> new_vert_indices;
 	std::vector<RayMeshVertex> newverts;
@@ -712,7 +712,7 @@ void RayMesh::mergeVerticesWithSamePosAndNormal()
 
 	vertices = newverts;
 
-	conPrint("\tNew num vertices: " + toString(vertices.size()) + "");
+	conPrint("\tNew num vertices: " + toString((unsigned int)vertices.size()) + "");
 	conPrint("\tDone.");
 }
